History Logic refactoring
