+++ app/src/main/java/com/example/nvdovin/weatherapp/presentation/history/grid/GridHistoryView.java
@@ -1,8 +1,6 @@
 package com.example.nvdovin.weatherapp.presentation.history.grid;
 
 import android.content.Context;
-import android.content.Intent;
-import android.os.Bundle;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -13,7 +11,6 @@ import android.widget.TextView;
 
 import com.example.nvdovin.weatherapp.R;
 import com.example.nvdovin.weatherapp.data.model.WeatherData;
-import com.example.nvdovin.weatherapp.presentation.details.DetailActivity;
 import com.example.nvdovin.weatherapp.presentation.history.grid.adapter.GridHistoryAdapter;
 
 import java.util.List;
@@ -23,22 +20,16 @@ import butterknife.ButterKnife;
 
 public class GridHistoryView {
 
-    private static final String DAY_TIMESTAMP = "TIMESTAMP_KEY";
-    private static final String CITY_ID = "CITY_ID_KEY";
-    private static final String DETAIL_BUNDLE = "DETAIL_BUNDLE";
-
     @BindView(R.id.weather_grid_view)
     GridView weatherGridView;
     @BindView(R.id.no_data)
     TextView noData;
 
-    private Context context;
     private View view;
     private GridHistoryAdapter gridHistoryAdapter;
 
     public GridHistoryView(Context context, GridHistoryAdapter gridHistoryAdapter) {
 
-        this.context = context;
         this.gridHistoryAdapter = gridHistoryAdapter;
 
         FrameLayout frameLayout = new FrameLayout(context);
@@ -61,21 +52,10 @@ public class GridHistoryView {
         return view;
     }
 
-
-    public void displayHistory(final List<WeatherData> updatedWeatherDataList) {
+    public void displayHistory(final List<WeatherData> updatedWeatherDataList, AdapterView.OnItemClickListener itemClickListener) {
         gridHistoryAdapter.swap(updatedWeatherDataList);
         noData.setVisibility(updatedWeatherDataList.isEmpty() ? View.VISIBLE : View.GONE);
-        weatherGridView.setOnItemClickListener(new AdapterView.OnItemClickListener() {
-            @Override
-            public void onItemClick(AdapterView<?> parent, View view, int position, long id) {
-                Bundle bundle = new Bundle();
-                bundle.putLong(CITY_ID, updatedWeatherDataList.get(position).getCityId());
-                bundle.putLong(DAY_TIMESTAMP, updatedWeatherDataList.get(position).getDt());
-                Intent intent = new Intent(context, DetailActivity.class);
-                intent.putExtra(DETAIL_BUNDLE, bundle);
-                context.startActivity(intent);
-            }
-        });
+        weatherGridView.setOnItemClickListener(itemClickListener);
     }
 
 }
